+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.1] - 2023-04-29
+### Added
+- Adds Covariance Matrix adaptation (CMA) algorithm
+- Adds a damping term for SVD inverse operations
+### Changed
+- Fixes a potential segfault issue with logging interface
+- Fixes a header install path CMake issue
+
 ## [0.4.0] - 2023-01-30
 ### Added
 - Adds metrics for controllers and estimators